实现了回复相关的异常处理

+++ src/main/java/com/lingerlin/community/community/dto/CommentDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
  */
 @Data
 public class CommentDTO {
-    private Long parentId;
+    private Integer parentId;
     private String content;
     private Integer type;
 }
