+++ src/main/java/com/lingerlin/community/community/model/Comment.java
@@ -7,8 +7,8 @@ import lombok.Data;
  */
 @Data
 public class Comment {
-    private Long id;
-    private Long parentId;
+    private Integer id;
+    private Integer parentId;
     private Integer type;
     private Integer commentator;
     private Long gmtCreate;
